requirements.txt
requests==2.10.0
beautifulsoup4==4.4.1
python-dateutil==2.5.3
python-slugify==1.2.0
#ipdb==0.10.0
#ipython==4.2.0
